chore(utils): extract env loader helpers from ns-proxy
ssh-gateway will share the same RCP_ env loading patterns (sock path,
positive duration, log level). Move EnvInt / EnvPositiveDuration /
EnvSockPath / EnvLogLevel into a shared internal/utils package and
re-point cmd/ns-proxy/config.go to use them.

No behavior change; ns-proxy tests still pass.
