+++ cmd/ns-proxy/config.go
@@ -2,10 +2,9 @@ package main
 
 import (
 	"fmt"
-	"path/filepath"
-	"strconv"
-	"strings"
 	"time"
+
+	"github.com/KHU-RETURN/rcp-server/internal/utils"
 )
 
 type Config struct {
@@ -24,26 +23,26 @@ func LoadConfig(getenv func(string) string) (*Config, error) {
 	if err != nil {
 		return nil, fmt.Errorf("RCP_NS_PROXY_ALLOW_CIDR: %w", err)
 	}
-	maxConns, err := envInt(getenv, "RCP_NS_PROXY_MAX_CONNS", 1024)
+	maxConns, err := utils.EnvInt(getenv, "RCP_NS_PROXY_MAX_CONNS", 1024)
 	if err != nil {
 		return nil, err
 	}
 	if maxConns <= 0 {
 		return nil, fmt.Errorf("RCP_NS_PROXY_MAX_CONNS: must be positive, got %d", maxConns)
 	}
-	dialTimeout, err := envPositiveDuration(getenv, "RCP_NS_PROXY_DIAL_TIMEOUT", 5*time.Second)
+	dialTimeout, err := utils.EnvPositiveDuration(getenv, "RCP_NS_PROXY_DIAL_TIMEOUT", 5*time.Second)
 	if err != nil {
 		return nil, err
 	}
-	grace, err := envPositiveDuration(getenv, "RCP_NS_PROXY_SHUTDOWN_GRACE", 30*time.Second)
+	grace, err := utils.EnvPositiveDuration(getenv, "RCP_NS_PROXY_SHUTDOWN_GRACE", 30*time.Second)
 	if err != nil {
 		return nil, err
 	}
-	sockPath, err := envSockPath(getenv, "RCP_NS_PROXY_SOCK", "/run/rcp/ns-proxy.sock")
+	sockPath, err := utils.EnvSockPath(getenv, "RCP_NS_PROXY_SOCK", "/run/rcp/ns-proxy.sock")
 	if err != nil {
 		return nil, err
 	}
-	logLevel, err := envLogLevel(getenv, "RCP_NS_PROXY_LOG_LEVEL", "info")
+	logLevel, err := utils.EnvLogLevel(getenv, "RCP_NS_PROXY_LOG_LEVEL", "info")
 	if err != nil {
 		return nil, err
 	}
@@ -55,68 +54,4 @@ func LoadConfig(getenv func(string) string) (*Config, error) {
 		ShutdownGrace: grace,
 		LogLevel:      logLevel,
 	}, nil
-}
-
-func envInt(get func(string) string, key string, def int) (int, error) {
-	v := get(key)
-	if v == "" {
-		return def, nil
-	}
-	n, err := strconv.Atoi(v)
-	if err != nil {
-		return 0, fmt.Errorf("%s: %w", key, err)
-	}
-	return n, nil
-}
-
-func envDuration(get func(string) string, key string, def time.Duration) (time.Duration, error) {
-	v := get(key)
-	if v == "" {
-		return def, nil
-	}
-	d, err := time.ParseDuration(v)
-	if err != nil {
-		return 0, fmt.Errorf("%s: %w", key, err)
-	}
-	return d, nil
-}
-
-func envPositiveDuration(get func(string) string, key string, def time.Duration) (time.Duration, error) {
-	d, err := envDuration(get, key, def)
-	if err != nil {
-		return 0, err
-	}
-	if d <= 0 {
-		return 0, fmt.Errorf("%s: must be > 0, got %v", key, d)
-	}
-	return d, nil
-}
-
-func envSockPath(get func(string) string, key, def string) (string, error) {
-	raw := get(key)
-	v := strings.TrimSpace(raw)
-	if raw != "" && v == "" {
-		// 값은 설정됐지만 공백만 있는 경우.
-		return "", fmt.Errorf("%s: must be an absolute path, got %q", key, raw)
-	}
-	if v == "" {
-		v = def
-	}
-	if !filepath.IsAbs(v) {
-		return "", fmt.Errorf("%s: must be an absolute path, got %q", key, v)
-	}
-	return v, nil
-}
-
-func envLogLevel(get func(string) string, key, def string) (string, error) {
-	v := strings.ToLower(strings.TrimSpace(get(key)))
-	if v == "" {
-		return def, nil
-	}
-	switch v {
-	case "debug", "info", "warn", "error":
-		return v, nil
-	default:
-		return "", fmt.Errorf("%s: invalid log level %q (allowed: debug, info, warn, error)", key, v)
-	}
 }
